Added a tutorial on importing PhenoCam data using R.
This tutorial uses the phenocamr and ggplot2 packages to download
PhenoCam GCC time series data, view time-series data, and calculate and
view seasonal transition dates.

The markdown tutorial was placed in the Importing Data section of the
geospatial workbook, with corresponding modifications to the landing
page and assets, and the R markdown tutorial was placed in the tutorials
section.